+++ src/core/pipeline.ts
@@ -13,7 +13,13 @@ import {
     Recognizer,
     Token,
 } from 'antlr4'
-import { isDebug, isDev, localAppEnv, localNodeEnv } from '../config/env'
+import {
+    isDebug,
+    isDev,
+    isProdEnv,
+    localAppEnv,
+    localNodeEnv,
+} from '../config/env'
 import YiniLexer from '../grammar/generated/YiniLexer'
 import YiniParser, { YiniContext } from '../grammar/generated/YiniParser'
 import {
@@ -77,6 +83,11 @@ const parsePossibleStartCol = (
         } else {
             msgHint = 'Thrown value:' + JSON.stringify(err)
         }
+
+        if (isProdEnv()) {
+            return 0
+        }
+
         errorHandler.pushOrBail(
             null,
             'Internal-Error',
@@ -190,7 +201,7 @@ export const runPipeline = (
 ): ParsedObject | YiniParseResult => {
     debugPrint()
     debugPrint('-> Entered parseMain(..) in parseEntry')
-    debugPrint('           isStrict mode = ' + coreOptions.isStrict)
+    debugPrint('    isStrict initialMode = ' + coreOptions.rules.initialMode)
     debugPrint('         bailSensitivity = ' + coreOptions.bailSensitivity)
     debugPrint('           isIncludeMeta = ' + coreOptions.isIncludeMeta)
     debugPrint('       isWithDiagnostics = ' + coreOptions.isWithDiagnostics)
@@ -201,11 +212,13 @@ export const runPipeline = (
     debugPrint(
         'isAvoidWarningsInConsole = ' + coreOptions.isAvoidWarningsInConsole,
     )
-    debugPrint('    requireDocTerminator = ' + coreOptions.requireDocTerminator)
+    debugPrint('          onDuplicateKey = ' + coreOptions.rules.onDuplicateKey)
     debugPrint(
-        '   treatEmptyValueAsNull = ' + coreOptions.treatEmptyValueAsNull,
+        '    requireDocTerminator = ' + coreOptions.rules.requireDocTerminator,
     )
-    debugPrint('          onDuplicateKey = ' + coreOptions.onDuplicateKey)
+    debugPrint(
+        '   treatEmptyValueAsNull = ' + coreOptions.rules.treatEmptyValueAsNull,
+    )
     debugPrint()
     debugPrint('  runtimeInfo.sourceType = ' + runtimeInfo.sourceType)
     debugPrint('    runtimeInfo.fileName = ' + runtimeInfo.fileName)
@@ -400,12 +413,12 @@ export const runPipeline = (
     isDebug() && console.debug(finalJSResult)
     debugPrint()
 
-    if (coreOptions.isStrict) {
+    if (coreOptions.rules.initialMode === 'strict') {
         // Note, after pushing processing may continue or exit, depending on the error and/or the bail threshold.
         errorHandler.pushOrBail(
             null,
             'Syntax-Warning',
-            'Warning: Strict mode is not yet fully implemented.',
+            'Warning: Strict initialMode is not yet fully implemented.',
             'Some validation rules may still be missing or incomplete.',
         )
     } else {
@@ -432,7 +445,7 @@ export const runPipeline = (
 
     debugPrint('getNumOfErrors(): ' + errorHandler.getNumOfErrors())
     if (errorHandler.getNumOfErrors()) {
-        console.log()
+        // console.log()
         console.log(
             'Parsing is complete, but some problems were detected. Please see the errors above for details.',
         )