+++ src/core/runtime.ts
@@ -10,13 +10,13 @@ import {
     TBailSensitivityLevel,
     TParserMode,
 } from './internalTypes'
+import { getDefaultUserOptions } from './options/defaultParserOptions'
 import { mapFailLevelToBail } from './options/failLevel'
 import {
     inferModeFromArgs,
     isOptionsObjectForm,
     toCoreOptions,
-} from './options/normalizeOptions'
-import { getDefaultOptions } from './options/parserOptionsConstants'
+} from './options/optionsFunctions'
 import { runPipeline } from './pipeline'
 
 /**
@@ -90,7 +90,7 @@ export class YiniRuntime {
         }
 
         const mode: TParserMode = inferModeFromArgs(arg2)
-        const defaultOptions = getDefaultOptions(mode)
+        const defaultOptions = getDefaultUserOptions(mode)
 
         // Normalize to a fully-required options object.
         let userOpts: Required<ParseOptions>
@@ -136,7 +136,7 @@ export class YiniRuntime {
             userOpts.failLevel,
         )
 
-        const coreOpts: IParseCoreOptions = toCoreOptions(level, userOpts)
+        const coreOpts: IParseCoreOptions = toCoreOptions(userOpts, level)
 
         debugPrint()
         debugPrint('==== Call runPipeline(..) ==========================')
@@ -200,7 +200,7 @@ export class YiniRuntime {
         }
 
         const mode: TParserMode = inferModeFromArgs(arg2)
-        const defaultOptions = getDefaultOptions(mode)
+        const defaultOptions = getDefaultUserOptions(mode)
 
         // Normalize to a fully-required options object.
         let userOpts: Required<ParseOptions>